Fix chunk border for 3d chunk

+++ Assets/Chunk/ChunkData.cs
@@ -2,10 +2,10 @@ using UnityEngine;
 
 public class ChunkData : MonoBehaviour
 {
-    private float chunkSize;
-    public float ChunkSize => chunkSize;
+    private int chunkSize;
+    public int ChunkSize => chunkSize;
 
-    public void Initialize(float chunkSize)
+    public void Initialize(int chunkSize)
     {
         this.chunkSize = chunkSize;
     }
